+++ docs/extra.js
@@ -34,6 +34,16 @@ $( document ).ready(function() {
     window.location.replace(url_new);
   }
 
+  $('#sidebar').prepend(
+    '<div id="r4ds">' +
+    '  <a href="https://r4ds.had.co.nz/">' +
+    '    <img src="../../AMR/cover_r4ds.png" width="25%">' +
+    '  </a>' +
+    '  <p>Learn R reading this great book!</p>' +
+    '  <p>Or read it free online: <a href="https://r4ds.had.co.nz/" target="_blank">r4ds.co.nz</a>.</p>' +
+    '  <hr>' +
+    '</div>');
+
   $('footer').html(
     '<div>' +
       '<p>' + $('footer .copyright p').html().replace(